move whisk state transition code to its own module
